+++ packages/talker_dio_logger/example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:talker_dio_logger/talker_dio_logger.dart';
+import 'package:talker_flutter/talker_flutter.dart';
 
 void main() {
   runApp(const MyApp());
@@ -18,9 +19,14 @@ class _MyAppState extends State<MyApp> {
 
   @override
   void initState() {
+    // Disable flutter logs cut by the console
+    final talker = Talker(loggerOutput: debugPrint);
+    // Disable flutter logs cut by the console
+
     _dio = Dio();
     _dio.interceptors.add(
       TalkerDioLogger(
+        talker: talker,
         settings: const TalkerDioLoggerSettings(
           printRequestHeaders: true,
           printResponseHeaders: true,
@@ -61,9 +67,30 @@ class _MyAppState extends State<MyApp> {
               ),
               ElevatedButton(
                 onPressed: () {
+                  _dio.put('https://dummyjson.com/products/1');
+                },
+                child: const Text('PUT products request'),
+              ),
+              ElevatedButton(
+                onPressed: () {
+                  _dio.post('https://dummyjson.com/products/add');
+                },
+                child: const Text('POST products request'),
+              ),
+              ElevatedButton(
+                onPressed: () {
+                  _dio.delete('https://dummyjson.com/products/1');
+                },
+                child: const Text('DELETE products request'),
+              ),
+              ElevatedButton(
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all(Colors.red),
+                ),
+                onPressed: () {
                   _dio.get('htt://jsonplaceholder.typicode.com/todos');
                 },
-                child: const Text('Invalid request with exception'),
+                child: const Text('Invalid GET request with error'),
               ),
             ],
           ),