+++ packages/talker_dio_logger/lib/talker_dio_logger_settings.dart
@@ -1,3 +1,5 @@
+import 'package:talker/talker.dart';
+
 /// [TalkerDioLogger] settings and customization
 class TalkerDioLoggerSettings {
   const TalkerDioLoggerSettings({
@@ -6,6 +8,9 @@ class TalkerDioLoggerSettings {
     this.printResponseMessage = true,
     this.printRequestData = true,
     this.printRequestHeaders = false,
+    this.requestPen,
+    this.responsePen,
+    this.errorPen,
   });
 
   /// Print [response.data] if true
@@ -22,4 +27,37 @@ class TalkerDioLoggerSettings {
 
   /// Print [request.headers] if true
   final bool printRequestHeaders;
+
+  /// Field to set custom http request console logs color
+  ///```
+  ///// Red color
+  ///final redPen = AnsiPen()..red();
+  ///
+  ///// Blue color
+  ///final redPen = AnsiPen()..blue();
+  ///```
+  /// More details in [AnsiPen] docs
+  final AnsiPen? requestPen;
+
+  /// Field to set custom http response console logs color
+  ///```
+  ///// Red color
+  ///final redPen = AnsiPen()..red();
+  ///
+  ///// Blue color
+  ///final redPen = AnsiPen()..blue();
+  ///```
+  /// More details in [AnsiPen] docs
+  final AnsiPen? responsePen;
+
+  /// Field to set custom http error console logs color
+  ///```
+  ///// Red color
+  ///final redPen = AnsiPen()..red();
+  ///
+  ///// Blue color
+  ///final redPen = AnsiPen()..blue();
+  ///```
+  /// More details in [AnsiPen] docs
+  final AnsiPen? errorPen;
 }